DECRBY, HINCRBY, and INCRBY should all accept Long values

+++ src/main/scala/com/redis/HashOperations.scala
@@ -68,7 +68,7 @@ trait HashOperations {
       }
     }
 
-  def hincrby(key: Any, field: Any, value: Int)(implicit format: Format): Option[Long] =
+  def hincrby(key: Any, field: Any, value: Long)(implicit format: Format): Option[Long] =
     send("HINCRBY", List(key, field, value))(asLong)
 
   def hincrbyfloat(key: Any, field: Any, value: Float)(implicit format: Format): Option[Float] =
